Move model definitions to subfolder
