+++ src/test_cadet_core/crystallization_partII.py
@@ -28,7 +28,7 @@ from cadet import Cadet
 from cadetrdm import ProjectRepo
 
 import utility.convergence as convergence
-import settings_crystallization
+from benchmark_models import settings_crystallization
 
 
 # %% Helper functions
